refactor import structure and components file
