+++ src/components/index.js
@@ -1,0 +1,4 @@
+import Visualiser from "./visulaiser";
+
+export { Visualiser };
+export default Visualiser;
